requirements.txt
pandas
numpy

streamlit
requests
streamlit_lottie
gensim
google-cloud-storage
